do not invoke history on Mac/Darwin (closes #77)

+++ R/rite.r
@@ -13,7 +13,8 @@ rite <- function(filename=NULL, catchOutput=FALSE, evalenv=.GlobalEnv,
         echorun <- tclVar(1)
     else
         echorun <- tclVar(0)
-    addtohistory <- tclVar(1)
+    if(!Sys.info()['sysname'] == "Darwin")
+        addtohistory <- tclVar(1)
     everWarn <- tclVar(1)
     # optionally setup evaluation environment
     if(is.null(evalenv)){
@@ -319,11 +320,13 @@ rite <- function(filename=NULL, catchOutput=FALSE, evalenv=.GlobalEnv,
             }
         }
     }
-    loadHistory <- function(){
-        tmphistory <- tempfile("Rrawhist")
-        savehistory(tmphistory)
-        loadScript(tmphistory)
-        unlink(tmphistory)
+    if(!Sys.info()['sysname'] == "Darwin") {        
+        loadHistory <- function(){
+            tmphistory <- tempfile("Rrawhist")
+            savehistory(tmphistory)
+            loadScript(tmphistory)
+            unlink(tmphistory)
+        }
     }
     saveScript <- function(){
         if(is.null(filename) || !length(filename) || filename=="")
@@ -565,7 +568,7 @@ rite <- function(filename=NULL, catchOutput=FALSE, evalenv=.GlobalEnv,
         # a restart to 'discontinue' source(.)-ing
         discontinue = function(){invisible()}
         )
-        if(as.logical(as.numeric(tclvalue(addtohistory)))){
+        if(!Sys.info()['sysname'] == "Darwin" && as.logical(as.numeric(tclvalue(addtohistory)))){
             tmphistory <- tempfile()
             savehistory(tmphistory)
             histcon <- file(tmphistory, open="a")
@@ -1000,8 +1003,10 @@ rite <- function(filename=NULL, catchOutput=FALSE, evalenv=.GlobalEnv,
         tkadd(menuFile, "command", label="New Script", command=newScript, underline = 0)
         tkadd(menuFile, "command", label="Load Script",
             command=function() loadScript(locals=TRUE), underline = 0)
-        tkadd(menuFile, "command", label="Load Command History",
-            command=loadHistory)
+        if(!Sys.info()['sysname'] == "Darwin") {
+            tkadd(menuFile, "command", label="Load Command History",
+                command=loadHistory)
+        }
         tkadd(menuFile, "command", label="Save Script", command=saveScript, underline = 0)
         tkadd(menuFile, "command", label="SaveAs Script", command=saveAsScript, underline = 1)
         tkadd(menuFile, "command", label="Append Script",
@@ -1066,7 +1071,8 @@ rite <- function(filename=NULL, catchOutput=FALSE, evalenv=.GlobalEnv,
         tkadd(menuRun, "separator")
         tkadd(menuRun, 'checkbutton', label='Echo code?', onvalue=1L, variable=echorun)
         tkadd(menuRun, 'checkbutton', label='Wait on warnings?', onvalue=1L, variable=everWarn)
-        tkadd(menuRun, 'checkbutton', label='Add commands to history?', onvalue=1L, variable=addtohistory)
+        if(!Sys.info()['sysname'] == "Darwin")
+            tkadd(menuRun, 'checkbutton', label='Add commands to history?', onvalue=1L, variable=addtohistory)
         tkadd(menuRun, "separator")
         if(catchOutput){
             tkadd(menuRun, "command", label="List all objects", command=function()
